# GeneratedApiClient/src/StudentSurveySystem.ApiClient/Model/UserRole.cs
/* 
 * Student survey system API
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System.Runtime.Serialization;
using Newtonsoft.Json;
using Newtonsoft.Json.Converters;

namespace StudentSurveySystem.ApiClient.Model
{
    /// <summary>
    /// Defines UserRole
    /// </summary>
    [JsonConverter(typeof(StringEnumConverter))]
        public enum UserRole
    {
        /// <summary>
        /// Enum Student for value: Student
        /// </summary>
        [EnumMember(Value = "Student")]
        Student = 1,
        /// <summary>
        /// Enum Lecturer for value: Lecturer
        /// </summary>
        [EnumMember(Value = "Lecturer")]
        Lecturer = 2,
        /// <summary>
        /// Enum Admin for value: Admin
        /// </summary>
        [EnumMember(Value = "Admin")]
        Admin = 3    }
}